+++ modules/aggregation-layers/src/index.ts
@@ -36,7 +36,9 @@ export {default as _BinSorter} from './utils/bin-sorter';
 
 // types
 export type {ContourLayerProps} from './contour-layer/contour-layer';
+export type {HeatmapLayerProps} from './heatmap-layer/heatmap-layer';
 export type {HexagonLayerProps} from './hexagon-layer/hexagon-layer';
 export type {CPUGridLayerProps} from './cpu-grid-layer/cpu-grid-layer';
 export type {GridLayerProps} from './grid-layer/grid-layer';
 export type {GPUGridLayerProps} from './gpu-grid-layer/gpu-grid-layer';
+export type {ScreenGridLayerProps} from './screen-grid-layer/screen-grid-layer';