+++ modules/core/src/effects/lighting/camera-light.ts
@@ -1,3 +1,4 @@
+/* eslint-disable camelcase */
 import {PointLight} from './point-light';
 import {getUniformsFromViewport} from '../../shaderlib/project/viewport-uniforms';
 import type Layer from '../../lib/layer';
@@ -7,7 +8,6 @@ export default class CameraLight extends PointLight {
     const {projectedLight} = this;
     const viewport = layer.context.viewport;
     const {coordinateSystem, coordinateOrigin, modelMatrix} = layer.props;
-    // @ts-expect-error
     const {project_uCameraPosition} = getUniformsFromViewport({
       viewport,
       modelMatrix,