+++ modules/core/src/index.ts
@@ -119,9 +119,13 @@ export type {
   Unit,
   Position,
   Color,
-  Texture
+  Texture,
+  Material
 } from './types/layer-props';
+export type {FilterContext} from './passes/layers-pass';
 export type {PickingInfo, GetPickingInfoParams} from './lib/picking/pick-info';
-export type {ConstructorOf as _ConstructorOf} from './types/types';
+export type {ConstructorOf as _ConstructorOf, ShaderModule as _ShaderModule} from './types/types';
 export type {BinaryAttribute} from './lib/attribute/attribute';
 export type {Effect} from './lib/effect';
+export type {PickingUniforms, ProjectUniforms} from './shaderlib';
+export type {DefaultProps} from './lifecycle/prop-types';