+++ modules/geo-layers/src/quadkey-layer/quadkey-layer.ts
@@ -1,8 +1,8 @@
-import {AccessorFunction} from '@deck.gl/core';
+import {AccessorFunction, DefaultProps} from '@deck.gl/core';
 import GeoCellLayer, {GeoCellLayerProps} from '../geo-cell-layer/GeoCellLayer';
 import {getQuadkeyPolygon} from './quadkey-utils';
 
-const defaultProps = {
+const defaultProps: DefaultProps<QuadkeyLayerProps> = {
   getQuadkey: {type: 'accessor', value: d => d.quadkey}
 };
 
@@ -10,7 +10,7 @@ const defaultProps = {
 export type QuadkeyLayerProps<DataT = any> = _QuadkeyLayerProps<DataT> & GeoCellLayerProps<DataT>;
 
 /** Properties added by QuadkeyLayer. */
-type _QuadkeyLayerProps<DataT = any> = {
+type _QuadkeyLayerProps<DataT> = {
   /**
    * Called for each data object to retrieve the quadkey string identifier.
    *
@@ -21,10 +21,10 @@ type _QuadkeyLayerProps<DataT = any> = {
 
 export default class QuadkeyLayer<DataT = any, ExtraProps = {}> extends GeoCellLayer<
   DataT,
-  Required<_QuadkeyLayerProps> & ExtraProps
+  Required<_QuadkeyLayerProps<DataT>> & ExtraProps
 > {
   static layerName = 'QuadkeyLayer';
-  static defaultProps: any = defaultProps;
+  static defaultProps = defaultProps;
 
   indexToBounds(): Partial<GeoCellLayer['props']> | null {
     const {data, getQuadkey} = this.props;