+++ modules/layers/src/index.ts
@@ -54,4 +54,3 @@ export type {PointCloudLayerProps} from './point-cloud-layer/point-cloud-layer';
 export type {TextBackgroundLayerProps} from './text-layer/text-background-layer/text-background-layer';
 export type {PathLayerProps} from './path-layer/path-layer';
 export type {SolidPolygonLayerProps} from './solid-polygon-layer/solid-polygon-layer';
-export type {MaterialProps as _MaterialProps} from './types';