+++ modules/mapbox/src/mapbox-overlay.ts
@@ -5,27 +5,49 @@ import type {Map, IControl, MapMouseEvent} from 'mapbox-gl';
 import type {MjolnirGestureEvent, MjolnirPointerEvent} from 'mjolnir.js';
 import type {DeckProps} from '@deck.gl/core';
 
+import {resolveLayers} from './resolve-layers';
+
+export type MapboxOverlayProps = Omit<
+  DeckProps,
+  | 'width'
+  | 'height'
+  | 'gl'
+  | 'parent'
+  | 'canvas'
+  | '_customRender'
+  | 'viewState'
+  | 'initialViewState'
+  | 'controller'
+>;
+
 /**
  * Implements Mapbox [IControl](https://docs.mapbox.com/mapbox-gl-js/api/markers/#icontrol) interface
  * Renders deck.gl layers over the base map and automatically synchronizes with the map's camera
  */
 export default class MapboxOverlay implements IControl {
-  private _props: DeckProps;
+  private _props: MapboxOverlayProps;
   private _deck?: Deck;
   private _map?: Map;
   private _container?: HTMLDivElement;
+  private _interleaved: boolean;
 
-  constructor(props: DeckProps) {
-    this._props = {...props};
+  constructor(
+    props: MapboxOverlayProps & {
+      interleaved?: boolean;
+    }
+  ) {
+    const {interleaved = false, ...otherProps} = props;
+    this._interleaved = interleaved;
+    this._props = otherProps;
   }
 
   /** Update (partial) props of the underlying Deck instance. */
-  setProps(props: DeckProps): void {
-    Object.assign(this._props, props);
-
-    if ('viewState' in this._props) {
-      delete this._props.viewState;
+  setProps(props: MapboxOverlayProps): void {
+    if (this._interleaved && props.layers) {
+      resolveLayers(this._map, this._deck, this._props.layers, props.layers);
     }
+
+    Object.assign(this._props, props);
 
     if (this._deck) {
       this._deck.setProps(this._props);
@@ -35,7 +57,10 @@ export default class MapboxOverlay implements IControl {
   /** Called when the control is added to a map */
   onAdd(map: Map): HTMLDivElement {
     this._map = map;
+    return this._interleaved ? this._onAddInterleaved(map) : this._onAddOverlaid(map);
+  }
 
+  private _onAddOverlaid(map: Map): HTMLDivElement {
     /* global document */
     const container = document.createElement('div');
     Object.assign(container.style, {
@@ -63,23 +88,49 @@ export default class MapboxOverlay implements IControl {
     return container;
   }
 
+  private _onAddInterleaved(map: Map): HTMLDivElement {
+    this._deck = new Deck({
+      ...this._props,
+      // @ts-ignore non-public map property
+      gl: map.painter.context.gl
+    });
+
+    map.on('styledata', this._handleStyleChange);
+    resolveLayers(map, this._deck, [], this._props.layers);
+
+    return document.createElement('div');
+  }
+
   /** Called when the control is removed from a map */
   onRemove(): void {
     const map = this._map;
 
     if (map) {
-      map.off('resize', this._updateContainerSize);
-      map.off('render', this._updateViewState);
-      map.off('mousemove', this._handleMouseEvent);
-      map.off('mouseout', this._handleMouseEvent);
-      map.off('click', this._handleMouseEvent);
-      map.off('dblclick', this._handleMouseEvent);
+      if (this._interleaved) {
+        this._onRemoveInterleaved(map);
+      } else {
+        this._onRemoveOverlaid(map);
+      }
     }
 
     this._deck?.finalize();
     this._deck = undefined;
     this._map = undefined;
     this._container = undefined;
+  }
+
+  private _onRemoveOverlaid(map: Map): void {
+    map.off('resize', this._updateContainerSize);
+    map.off('render', this._updateViewState);
+    map.off('mousemove', this._handleMouseEvent);
+    map.off('mouseout', this._handleMouseEvent);
+    map.off('click', this._handleMouseEvent);
+    map.off('dblclick', this._handleMouseEvent);
+  }
+
+  private _onRemoveInterleaved(map: Map): void {
+    map.off('styledata', this._handleStyleChange);
+    resolveLayers(map, this._deck, this._props.layers, []);
   }
 
   getDefaultPosition() {
@@ -112,6 +163,10 @@ export default class MapboxOverlay implements IControl {
       this._map.removeControl(this);
     }
   }
+
+  private _handleStyleChange = () => {
+    resolveLayers(this._map, this._deck, this._props.layers, this._props.layers);
+  };
 
   private _updateContainerSize = () => {
     if (this._map && this._container) {