+++ test/modules/carto/layers/h3-tileset-2d.spec.js
@@ -12,14 +12,17 @@ test('H3Tileset2D', async t => {
     height: 200
   });
 
-  const indices = tileset.getTileIndices({viewport});
+  const indices = tileset
+    .getTileIndices({viewport})
+    // Sort for reliable test output
+    .sort((a, b) => parseInt(a.i, 16) - parseInt(b.i, 16));
   t.deepEqual(
     indices,
     [
+      {i: '8274effffffffff'},
+      {i: '827547fffffffff'},
       {i: '82754ffffffffff'},
       {i: '82755ffffffffff'},
-      {i: '827547fffffffff'},
-      {i: '8274effffffffff'},
       {i: '82756ffffffffff'}
     ],
     'indices in viewport'